commens for ETL steps added

+++ postgres/covidvspoverty.sql
@@ -2,13 +2,13 @@ CREATE TABLE covid_data (
 fips_code TEXT PRIMARY KEY,
 state_name TEXT,
 county_name TEXT,
-lat INT,
-lon INT,
+lat FLOAT,
+lon FLOAT,
 total_population INT,
 confirmed INT,
-confirmed_per_100000 INT,
+confirmed_per_100000 FLOAT,
 deaths INT,
-deaths_per_100000 INT
+deaths_per_100000 FLOAT
 );
 
 CREATE TABLE poverty_data (
